+++ src/mranderson/move.clj
@@ -6,7 +6,8 @@
 ;; agreeing to be bound by the terms of this license. You must not
 ;; remove this notice, or any other, from this software.
 
-(ns ^{:author "Stuart Sierra, Benedek Fazekas"
+(ns ^{:no-doc true
+      :author "Stuart Sierra, Benedek Fazekas"
       :doc    "Refactoring tool to move a Clojure namespace from one name/file to
   another, and update all references to that namespace in your other
   Clojure source files.